rename(@/constants): cart저장 키를 localStorageKeys.ts에서 관리 (#101)
cart정보 로컬스토리지 저장 key 문자열 상수를 localStorageKeys.ts에서 관리하도록 변경했습니다.
