--- src/constants/cart.ts
@@ -1,3 +1,0 @@
-const CART_STORAGE_KEY = "cart";
-
-export default CART_STORAGE_KEY;
