+++ src/constants/localstorageKeys.ts
@@ -1,1 +1,2 @@
 export const LOGIN_USER_INFO_KEY = "loggedInUser";
+export const CART_STORAGE_KEY = "cart";
